sc-qemu: Put the main loop in a separate thread.
The QEMU main loop is now called from a different thread than the
SystemC one. It allows for gracefully recovers from QEMU abort/exit.

Next step would be to switch back to the SystemC thread when doing QEMU
to SystemC upcalls.

Also reworked the way we call the main loop such that we allow QEMU to
block waiting for events. It provides a great speedup in vast majority of
cases. A limiter timer has been introduce to avoid being indefinitely
stuck in the main loop and have a chance to return to SystemC.